requirements.txt
openai
python-dotenv
typer[all]
pydantic
streamlit
agno
nest_asyncio
